feat(workout-session): implement workout session management with new components for session header, sets, and schema to enhance user experience and functionality in workout tracking. refactor(workout-builder): update stepper components to utilize new session management structure and improve code organization. fix(workout-session): correct import paths and component references to ensure proper functionality and maintainability.
